% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countrycodes}
\alias{countrycodes}
\title{List of ISO 3166 country codes}
\format{The countrycodes data frame contains 125 observations on 2 variables.
\describe{
  \item{country_name}{name of country}
  \item{country_code}{ISO 3166 country codes of country}
}}
\usage{
countrycodes
}
\description{
This data provides a set of 125 countries that will work for functions in holidayplanneR package.
}
\keyword{datasets}
