% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidaylist.R
\name{holidaylist}
\alias{holidaylist}
\title{Generating a dataframe of all holidays in a year}
\usage{
holidaylist(country = c("US"), year = "2018",
  api_key = Sys.getenv("CALENDARINDEX_API_KEY"))
}
\arguments{
\item{country}{a single ISO 3166 country code or a list of ISO 3166 country codes from the 'countrycode' dataset in this package.}

\item{year}{the year of interest}

\item{api_key}{personal CalendarIndex API key}
}
\description{
This function will generate a dataframe containing a list of all holidays containing all the holidays that are taking place within a specified year in a specified country.
}
\details{
The dataframe will include the date of holiday, the country, the name of holiday, and the type of holiday (public, observance, optional, etc).
}
\examples{
holidaylist("US", "2018", Sys.getenv("CALENDARINDEX_API_KEY"))
holidaylist(c("US", "AD"), "2016", api_key = Sys.getenv("CALENDARINDEX_API_KEY"))
}
