% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isitholiday.R
\name{isitholiday}
\alias{isitholiday}
\title{Finding out if specified date is a holiday}
\usage{
isitholiday(dateofinterest = "2018-01-01", country = "US",
  api_key = Sys.getenv("CALENDARINDEX_API_KEY"))
}
\arguments{
\item{dateofinterest}{a single date in the format "YYYY-MM-DD".}

\item{country}{a single ISO 3166 country code from the 'countrycode' dataset in this package.}

\item{api_key}{personal CalendarIndex API key}
}
\description{
This function answers the question: Is the specified date a holiday in the specified country and year?
}
\details{
This function will return TRUE if the specified date is a holiday in the specified country and year, FALSE if the specified date is not a holiday in the specified country and year.
}
\examples{
isitholiday("2018-12-31", "US", Sys.getenv("CALENDARINDEX_API_KEY"))
isitholiday("2016-08-06", "AD", api_key = Sys.getenv("CALENDARINDEX_API_KEY"))
}
