% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longholidays.R
\name{longholidays}
\alias{longholidays}
\title{Identifying long blocks of holidays}
\usage{
longholidays(country = c("US"), year = "2018",
  api_key = Sys.getenv("CALENDARINDEX_API_KEY"))
}
\arguments{
\item{country}{a single ISO 3166 country code or a list of ISO 3166 country codes from the 'countrycode' dataset in this package.}

\item{year}{the year of interest}

\item{api_key}{personal CalendarIndex API key}
}
\description{
This function will identify the long holidays taking place in the specified country in the specified year, in a specified year, which are mainly due to public holidays on Fridays or Mondays.
}
\details{
The dataframe will include the specified year, the country, the start date and day of the long holiday, the end date and day of the long holiday, the total length of the long holiday. Only public holidays are considered (observances and optional holidays are excluded).
}
\examples{
longholidays("US", "2018", Sys.getenv("CALENDARINDEX_API_KEY"))
longholidays(c("US", "AD"), "2016", api_key = Sys.getenv("CALENDARINDEX_API_KEY"))
}
